Move metadata decorators to their own module
* OSBS-7828
